+++ onem2mlib/exceptions.py
@@ -13,12 +13,15 @@ This sub-module defines the exceptions used in the onem2mlib module.
 
 
 class OneM2MLibError(Exception):
-    """Base class for exceptions in this module."""
+    """
+    Base class for exceptions in this module.
+    """
     pass
 
 
 class CSEOperationError(OneM2MLibError):
-    """Exception raised for errors when invoking operations on the CSE.
+    """
+    Exception raised for errors when invoking operations on the CSE.
     """
 
     def __init__(self, message):
@@ -27,7 +30,8 @@ class CSEOperationError(OneM2MLibError):
 
 
 class AuthenticationError(OneM2MLibError):
-    """Exception raised for errors regarding authorization. 
+    """
+    Exception raised for errors regarding authorization. 
     """
 
     def __init__(self, message):
@@ -36,7 +40,8 @@ class AuthenticationError(OneM2MLibError):
 
 
 class ParameterError(OneM2MLibError):
-    """Exception raised for errors in paramterization of resource classes. 
+    """
+    Exception raised for errors in paramterization of resource classes. 
     """
 
     def __init__(self, message):
@@ -46,9 +51,22 @@ class ParameterError(OneM2MLibError):
 
 
 class NotSupportedError(OneM2MLibError):
-    """Exception raised when accessing/modifying not supported features of a resource.
+    """
+    Exception raised when accessing/modifying not supported features of a resource.
     """
 
     def __init__(self, message):
         self.message = message
         """ Explanation of the error. """
+
+
+
+class EncodingError(OneM2MLibError):
+    """
+    Exception raised when receiving a wrong encoding.
+    """
+
+    def __init__(self, message):
+        self.message = message
+        """ Explanation of the error. """
+